backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
coverage==7.13.5
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
faiss-cpu==1.13.2
fastapi==0.135.2
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastapi-utils==0.8.0
fastar==0.9.0
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.10
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.22
langchain-openai==1.1.7
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
librt==0.8.1
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mirakuru==3.0.2
mpmath==1.3.0
multidict==6.7.1
mutagen==1.47.0
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.3
openai==2.30.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
pillow==12.1.1
pluggy==1.6.0
port-for==1.0.0
propcache==0.4.1
psutil==5.9.8
psycopg==3.3.3
psycopg2-binary==2.9.11
pyasn1==0.6.3
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.28
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentry-sdk==2.56.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlmodel==0.0.32
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.1.0
typer==0.24.1
typer-slim==0.24.0
types-aiofiles==25.1.0.20251011
types-passlib==1.7.7.20260211
types-pyasn1==0.6.0.20260319
types-python-jose==3.5.0.20250531
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0